+++ src/zugtyp.rs
@@ -1,10 +1,11 @@
 //! Alle Eigenschaften und bekannte Gleise für einen [Zugtyp].
 
-use std::marker::PhantomData;
+use std::{fmt::Debug, marker::PhantomData, time::Duration};
 
 use serde::{Deserialize, Serialize};
 
 use crate::{
+    eingeschränkt::NichtNegativ,
     gleis::{
         gerade::GeradeUnit,
         kreuzung::KreuzungUnit,
@@ -14,7 +15,7 @@ use crate::{
             s_kurve::SKurvenWeicheUnit,
         },
     },
-    steuerung::geschwindigkeit::BekannterLeiter,
+    steuerung::geschwindigkeit::{BekannterLeiter, Leiter},
     typen::mm::Spurweite,
 };
 
@@ -25,11 +26,13 @@ pub mod märklin;
 
 /// Spurweite, Leitervariante (als Phantomtyp) und alle bekannten Gleise
 #[derive(zugkontrolle_macros::Debug, zugkontrolle_macros::Clone)]
-pub struct Zugtyp<Leiter> {
+#[zugkontrolle_debug(<L as Leiter>::VerhältnisFahrspannungÜberspannung: Debug)]
+#[zugkontrolle_debug(<L as Leiter>::UmdrehenZeit: Debug)]
+pub struct Zugtyp<L: Leiter> {
     /// Der Name des Zugtyps.
     pub name: String,
     /// Die Leiter-Art des Zugtyps.
-    pub leiter: PhantomData<fn() -> Leiter>,
+    pub leiter: PhantomData<fn() -> L>,
     /// Spurweite
     pub spurweite: Spurweite,
     /// Alle unterstützten [Geraden](crate::gleis::gerade::Gerade).
@@ -46,11 +49,27 @@ pub struct Zugtyp<Leiter> {
     pub s_kurven_weichen: Vec<SKurvenWeicheUnit>,
     /// Alle unterstützten [Kreuzungen](crate::gleis::kreuzung::Kreuzung).
     pub kreuzungen: Vec<KreuzungUnit>,
+    /// Frequenz in Herz für den Pwm-Antrieb.
+    pub pwm_frequenz: NichtNegativ,
+    /// Verhältnis von maximaler Fahrspannung zu Überspannung zum Umdrehen.
+    pub verhältnis_fahrspannung_überspannung: <L as Leiter>::VerhältnisFahrspannungÜberspannung,
+    /// Zeit zum Anhalten vor dem Umdrehen.
+    pub stopp_zeit: Duration,
+    /// Zeit die zum Umdrehen verwendete Überspannung anliegt.
+    pub umdrehen_zeit: <L as Leiter>::UmdrehenZeit,
+    /// Zeit die Spannung an Weichen anliegt um diese zu schalten.
+    pub schalten_zeit: Duration,
 }
 
 /// Spurweite, Leitervariante (als Phantomtyp) und alle bekannten Gleise
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ZugtypSerialisiert {
+#[derive(zugkontrolle_macros::Debug, zugkontrolle_macros::Clone, Serialize, Deserialize)]
+#[zugkontrolle_debug(<L as Leiter>::VerhältnisFahrspannungÜberspannung: Debug)]
+#[zugkontrolle_debug(<L as Leiter>::UmdrehenZeit: Debug)]
+#[serde(bound(
+    serialize = "<L as Leiter>::VerhältnisFahrspannungÜberspannung: Serialize, <L as Leiter>::UmdrehenZeit: Serialize",
+    deserialize = "<L as Leiter>::VerhältnisFahrspannungÜberspannung: Deserialize<'de>, <L as Leiter>::UmdrehenZeit: Deserialize<'de>",
+))]
+pub struct ZugtypSerialisiert<L: Leiter> {
     /// Der Name des Zugtyps.
     pub name: String,
     /// Der [Name der Leiter-Art](BekannterLeiter::NAME) des Zugtyps.
@@ -71,16 +90,26 @@ pub struct ZugtypSerialisiert {
     pub s_kurven_weichen: Vec<SKurvenWeicheUnit>,
     /// Alle unterstützten [Kreuzungen](crate::gleis::kreuzung::Kreuzung).
     pub kreuzungen: Vec<KreuzungUnit>,
+    /// Frequenz in Herz für den Pwm-Antrieb.
+    pub pwm_frequenz: NichtNegativ,
+    /// Verhältnis von maximaler Fahrspannung zu Überspannung zum Umdrehen.
+    pub verhältnis_fahrspannung_überspannung: <L as Leiter>::VerhältnisFahrspannungÜberspannung,
+    /// Zeit zum Anhalten vor dem Umdrehen.
+    pub stopp_zeit: Duration,
+    /// Zeit die zum Umdrehen verwendete Überspannung anliegt.
+    pub umdrehen_zeit: <L as Leiter>::UmdrehenZeit,
+    /// Zeit die Spannung an Weichen anliegt um diese zu schalten.
+    pub schalten_zeit: Duration,
 }
 
 /// Der Leiter stimmt nicht mit dem Namen überein.
 #[derive(Debug, Clone)]
 pub struct FalscherLeiter(pub String);
 
-impl<Leiter: BekannterLeiter> TryFrom<ZugtypSerialisiert> for Zugtyp<Leiter> {
+impl<L: BekannterLeiter> TryFrom<ZugtypSerialisiert<L>> for Zugtyp<L> {
     type Error = FalscherLeiter;
 
-    fn try_from(serialisiert: ZugtypSerialisiert) -> Result<Self, Self::Error> {
+    fn try_from(serialisiert: ZugtypSerialisiert<L>) -> Result<Self, Self::Error> {
         let ZugtypSerialisiert {
             name,
             leiter,
@@ -92,8 +121,13 @@ impl<Leiter: BekannterLeiter> TryFrom<ZugtypSerialisiert> for Zugtyp<Leiter> {
             kurven_weichen,
             s_kurven_weichen,
             kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
         } = serialisiert;
-        let gesucht = Leiter::NAME.to_owned();
+        let gesucht = L::NAME.to_owned();
         if leiter != gesucht {
             return Err(FalscherLeiter(leiter));
         }
@@ -108,12 +142,17 @@ impl<Leiter: BekannterLeiter> TryFrom<ZugtypSerialisiert> for Zugtyp<Leiter> {
             kurven_weichen,
             s_kurven_weichen,
             kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
         })
     }
 }
 
-impl<Leiter: BekannterLeiter> From<Zugtyp<Leiter>> for ZugtypSerialisiert {
-    fn from(zugtyp: Zugtyp<Leiter>) -> Self {
+impl<L: BekannterLeiter> From<Zugtyp<L>> for ZugtypSerialisiert<L> {
+    fn from(zugtyp: Zugtyp<L>) -> Self {
         let Zugtyp {
             name,
             leiter: _,
@@ -125,8 +164,13 @@ impl<Leiter: BekannterLeiter> From<Zugtyp<Leiter>> for ZugtypSerialisiert {
             kurven_weichen,
             s_kurven_weichen,
             kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
         } = zugtyp;
-        let leiter = Leiter::NAME.to_owned();
+        let leiter = L::NAME.to_owned();
         ZugtypSerialisiert {
             name,
             leiter,
@@ -138,6 +182,11 @@ impl<Leiter: BekannterLeiter> From<Zugtyp<Leiter>> for ZugtypSerialisiert {
             kurven_weichen,
             s_kurven_weichen,
             kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
         }
     }
 }