+++ src/zugtyp/märklin.rs
@@ -3,9 +3,10 @@
 // HACK cargo check takes very long, this should reduce it until the lint is addressed
 #![allow(missing_docs)]
 
-use std::marker::PhantomData;
+use std::{marker::PhantomData, time::Duration};
 
 use crate::{
+    eingeschränkt::{NichtNegativ, NullBisEins},
     gleis::{
         gerade::{Gerade, GeradeUnit},
         kreuzung::{self, Kreuzung, KreuzungUnit},
@@ -63,6 +64,11 @@ impl Zugtyp<Mittelleiter> {
             kurven_weichen: vec![kurven_weiche_5140_links(), kurven_weiche_5140_rechts()],
             s_kurven_weichen: vec![],
             kreuzungen: vec![kreuzung_5128(), kreuzung_5207()],
+            pwm_frequenz: NichtNegativ::neu_unchecked(50.),
+            verhältnis_fahrspannung_überspannung: NullBisEins::neu_unchecked(16. / 25.),
+            stopp_zeit: Duration::from_millis(500),
+            umdrehen_zeit: Duration::from_millis(500),
+            schalten_zeit: Duration::from_millis(400),
         }
     }
 }